requirements.txt
pypdf
openpyxl
openai
plotly
